% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose-helpers.R
\name{add_pack_names}
\alias{add_pack_names}
\title{Add pack names to single exposures}
\usage{
add_pack_names(.single_exposures)
}
\arguments{
\item{.single_exposures}{List of \link[=single_exposure]{single exposures}.}
}
\description{
Function to add pack names to single exposures. Converts list of
\link[=single_exposure]{single exposures} to list of \link[=exposure]{exposures} without
validating.
}
\keyword{internal}
