% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{expose_single}
\alias{expose_single}
\title{Expose data to single rule pack}
\usage{
expose_single(.tbl, .pack, .rule_sep, .remove_obeyers, ...)
}
\arguments{
\item{.tbl}{Data frame of interest.}

\item{.pack}{Rule pack function.}

\item{.rule_sep}{Regular expression used as separator between column and
rule names in \link[=column-pack]{col packs} and \link[=cell-pack]{cell packs}.}

\item{.remove_obeyers}{Whether to remove elements which obey rules from
report.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The workhorse generic function for doing exposure. The result is
\link{single_exposure}.
}
\keyword{internal}
