% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inside_punct}
\alias{inside_punct}
\title{Inside punctuation regular expression}
\usage{
inside_punct(.x = "\\\\._\\\\.")
}
\arguments{
\item{.x}{Middle characters to be put between non alpha-numeric characters.}
}
\description{
Function to construct regular expression of form: 'non alpha-numeric
characters' + 'some characters' + 'non alpha-numeric characters'.
}
\examples{
inside_punct()
inside_punct("abc")

}
