% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyed-df.R
\name{keyed-df-dplyr}
\alias{keyed-df-dplyr}
\alias{select.keyed_df}
\alias{rename.keyed_df}
\alias{mutate.keyed_df}
\alias{summarise.keyed_df}
\alias{group_by.keyed_df}
\alias{ungroup.keyed_df}
\alias{rowwise.keyed_df}
\alias{arrange.keyed_df}
\alias{filter.keyed_df}
\alias{slice.keyed_df}
\title{Verbs from dplyr for keyed_df}
\usage{
\method{select}{keyed_df}(.tbl, ...)

\method{rename}{keyed_df}(.tbl, ...)

\method{mutate}{keyed_df}(.tbl, ...)

\method{summarise}{keyed_df}(.tbl, ...)

\method{group_by}{keyed_df}(.tbl, ..., add = FALSE)

\method{ungroup}{keyed_df}(.tbl, ...)

rowwise.keyed_df(.tbl)

\method{arrange}{keyed_df}(.tbl, ..., .by_group = FALSE)

\method{filter}{keyed_df}(.data, ...)

\method{slice}{keyed_df}(.tbl, ...)
}
\arguments{
\item{.tbl, .data}{A keyed object.}

\item{...}{Appropriate arguments for functions.}

\item{add}{Parameter for \link[dplyr:group_by]{dplyr::group_by}.}

\item{.by_group}{Parameter for \link[dplyr:arrange]{dplyr::arrange}.}
}
\description{
Defined methods for \link{dplyr} generic single table functions. They preserve
'keyed_df' class and 'keys' attribute (excluding \code{summarise} and its scoped
variants which remove them). Also these methods modify rows in keys according
to the rows modification in reference data.frame (if any).
}
\details{
\code{\link[dplyr:transmute]{dplyr::transmute()}} is supported implicitly with \code{\link[dplyr:mutate]{dplyr::mutate()}}
support.

\code{\link[dplyr:rowwise]{dplyr::rowwise()}} as for \code{dplyr} version 0.7.1 is not generic. Use
\code{rowwise.keyed_df} directly.

All \link[dplyr:scoped]{scoped} variants of present functions are also supported.
}
\examples{
mtcars \%>\% key_by(vs, am) \%>\% dplyr::mutate(gear = 1)

}
