% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{pack_info}
\alias{pack_info}
\alias{new_pack_info}
\title{Pack info}
\usage{
new_pack_info(.pack, .remove_obeyers)
}
\arguments{
\item{.pack}{\link[=rule-packs]{Rule pack}.}

\item{.remove_obeyers}{Value of \code{.remove_obeyers} argument of \code{\link[=expose]{expose()}} with
which \code{.pack} was applied.}
}
\description{
An S3 class \code{pack_info} to represent information about pack in \link[=single_exposure]{single exposure}. Its content is as in \link{packs_info} but without
column 'name'.
}
\keyword{internal}
