\name{Rule-class}
\Rdversion{1.1}
\docType{class}
\alias{Rule-class}

\title{Class \code{"Rule"}}
\description{
either a \code{SingleRule} object or \code{DoubleRule}
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "Rule" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("Rule")
}
\keyword{classes}
\keyword{internal}
