\name{vecORnull-class}
\Rdversion{1.1}
\docType{class}
\alias{vecORnull-class}

\title{Class \code{"vecORnull"}}
\description{
a class union for objects of class \code{NULL} or \code{vector} 
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "vecORnull" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("vecORnull")
}
\keyword{classes}
\keyword{internal}

