% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packs.R
\name{rule-packs}
\alias{rule-packs}
\alias{packs}
\alias{data_packs}
\alias{col_packs}
\alias{row_packs}
\alias{cell_packs}
\title{Create rule packs}
\usage{
data_packs(...)

col_packs(...)

row_packs(...)

cell_packs(...)
}
\arguments{
\item{...}{Rule functions which define pack. They can be in pure form or
inside a list (at any depth).}
}
\value{
\code{data_packs} returns a list of what should be \link[=data-pack]{data rule
packs}, \code{col_packs} - \link[=column-pack]{column rule packs},
\code{row_packs} - \link[=row-pack]{row rule packs}, \code{cell_packs} - \link[=cell-pack]{cell rule
packs}.
}
\description{
Functions for creating different kinds of rule packs. \strong{Rule} is a function
which converts object of interest (data, column, row, cell) to logical value
indicating whether this object satisfies certain condition. \strong{Rule pack} is
a function which combines several rules into one functional block. It takes a
data frame of interest and returns a data frame with only logical variables
and certain column naming scheme. Types of packs differ in interpretation of
their output.
}
\details{
These functions convert \code{...} to list, apply \code{rlang}'s
\link[rlang:squash]{squash()} and add appropriate classes. They are only
constructors and do not check for validity of certain pack. \strong{Note} that
it is allowed for elements of \code{...} to not have names: they will be computed
during exposure. However it is a good idea to manually name packs.
}
